+++ packages/types/src/codec/utils/encodeTypes.ts
@@ -103,12 +103,12 @@ function encodeTuple (typeDef: Pick<TypeDef, any>): string {
   })`;
 }
 
-function encodeUInt (typeDef: Pick<TypeDef, any>): string {
+function encodeUInt (typeDef: Pick<TypeDef, any>, type: 'Int' | 'UInt'): string {
   assert(typeDef.ext, 'Unable to encode VecFixed type');
 
   const { length } = typeDef.ext as TypeDefExtUInt;
 
-  return `UInt<${length}>`;
+  return `${type}<${length}>`;
 }
 
 function encodeVecFixed (typeDef: Pick<TypeDef, any>): string {
@@ -126,6 +126,7 @@ const encoders: Record<TypeDefInfo, (typeDef: TypeDef) => string> = {
   [TypeDefInfo.BTreeSet]: (typeDef: TypeDef): string => encodeWithParams(typeDef, 'BTreeSet'),
   [TypeDefInfo.Compact]: (typeDef: TypeDef): string => encodeWithParams(typeDef, 'Compact'),
   [TypeDefInfo.Enum]: (typeDef: TypeDef): string => encodeEnum(typeDef),
+  [TypeDefInfo.Int]: (typeDef: TypeDef): string => encodeUInt(typeDef, 'Int'),
   [TypeDefInfo.Linkage]: (typeDef: TypeDef): string => encodeWithParams(typeDef, 'Linkage'),
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   [TypeDefInfo.Null]: (typeDef: TypeDef): string => 'Null',
@@ -135,7 +136,7 @@ const encoders: Record<TypeDefInfo, (typeDef: TypeDef) => string> = {
   [TypeDefInfo.Set]: (typeDef: TypeDef): string => typeDef.type,
   [TypeDefInfo.Struct]: (typeDef: TypeDef): string => encodeStruct(typeDef),
   [TypeDefInfo.Tuple]: (typeDef: TypeDef): string => encodeTuple(typeDef),
-  [TypeDefInfo.UInt]: (typeDef: TypeDef): string => encodeUInt(typeDef),
+  [TypeDefInfo.UInt]: (typeDef: TypeDef): string => encodeUInt(typeDef, 'UInt'),
   [TypeDefInfo.Vec]: (typeDef: TypeDef): string => encodeWithParams(typeDef, 'Vec'),
   [TypeDefInfo.VecFixed]: (typeDef: TypeDef): string => encodeVecFixed(typeDef)
 };