+++ packages/types/src/interfaces/txpayment/types.ts
@@ -1,7 +1,7 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable @typescript-eslint/no-empty-interface */
 
-import { Fixed64 } from '@polkadot/types/primitive';
+import { Fixed64 } from '@polkadot/types/interfaces/runtime';
 
 /** @name Multiplier */
 export interface Multiplier extends Fixed64 {}